Join SVG plots. This function first populates each place in the first column, then rows in the seccond column

Join SVG plots. This function first populates each place in the first column, then rows in the seccond column

join_plots(..., nrows = length(list(...)), ncols = 1)

Arguments

...
mumltiple character vectors with SVG content

nrows
number of rows of plots in joint plot, default is set to number of plots

ncols
number of columns of plots in joint plot, default is set to 1

Value

Character vector with SVG content

Examples

  df <- data.frame(
   mon = month.abb[1:6],
   values = rnorm(6)
  )

join_plots(
  column_chart(df, x = 'mon', series = 'values'),
  column_chart(df, x = 'mon', series = 'values')
) %>% SVGrenderer()